+++ playwright/test-chandler-transitions-and-renumbering.spec.ts
@@ -50,17 +50,17 @@ test("test", async ({ page }) => {
     await globalThis.doSerialized();
   });
 
-  // Give it a 5 second fade.  Numeric inputs are flaky
-  await page.getByRole("row", { name: "c3" }).getByRole("spinbutton").fill("5");
+  // Give it a 10 second fade.  Numeric inputs are flaky
+  await page.getByRole("row", { name: "c3" }).getByRole("spinbutton").fill("10");
   await page.evaluate(async () => {
     await globalThis.doSerialized();
   });
-  await page.getByRole("row", { name: "c3" }).getByRole("spinbutton").fill("5");
+  await page.getByRole("row", { name: "c3" }).getByRole("spinbutton").fill("10");
   await page.evaluate(async () => {
     await globalThis.doSerialized();
   });
 
-  // go to c3
+  // select to c3
   await page.getByRole("cell", { name: "c3" }).click();
 
   await page.evaluate(async () => {
@@ -84,7 +84,7 @@ test("test", async ({ page }) => {
   await page.getByRole("button", { name: "Next 󰒭" }).click();
   await sleep(1000);
   await expect(page.getByTestId("sidebar-active-cue-name")).toContainText("c2");
-  await sleep(4500);
+  await sleep(4000);
   await expect(page.getByTestId("sidebar-active-cue-name")).toContainText("c3");
 
   // Ensure the tag point is actually fading in slowly
@@ -96,6 +96,8 @@ test("test", async ({ page }) => {
     .nth(1)
     .textContent();
   expect(Number.parseFloat(value1)).toBeGreaterThan(0);
+
+  // Yes this line is flaky. Sorry!
   expect(Number.parseFloat(value1)).toBeLessThan(255);
 
   await page.getByRole("link", { name: "󰓻 Tags" }).click();
@@ -106,10 +108,9 @@ test("test", async ({ page }) => {
     .nth(1)
     .textContent();
   expect(Number.parseFloat(value2)).toBeGreaterThan(0);
-  expect(Number.parseFloat(value2)).toBeLessThan(255);
-  expect(Number.parseFloat(value2)).toBeGreaterThan(Number.parseFloat(value1));
+  expect(Number.parseFloat(value2)).toBeGreaterThanOrEqual(Number.parseFloat(value1));
 
-  await sleep(5000);
+  await sleep(11000);
 
   await page.getByRole("link", { name: "󰓻 Tags" }).click();