+++ playwright/test_global_settings.spec.ts
@@ -9,16 +9,12 @@ test('test', async ({ page }) => {
     await page.getByRole('link', { name: '󰢻 System Settings' }).click();
 
 
-    await page.locator('input[name="apprise_target"]').click();
-    await page.locator('input[name="apprise_target"]').fill('gnome://');
-    await page.locator('form').filter({ hasText: 'This allows you to get' }).getByRole('button').click();
     await page.locator('input[name="url"]').click();
     await page.locator('input[name="url"]').fill('example.com');
     await page.locator('form').filter({ hasText: 'URL to redirect / to' }).getByRole('button').click();
 
     await page.getByRole('link', { name: '󰢻 Tools' }).click();
     await page.getByRole('link', { name: 'System Settings' }).click();
-    await expect(page.locator('input[name="apprise_target"]')).toHaveValue('gnome://');
     await expect(page.locator('input[name="url"]')).toHaveValue('example.com');
    
     await logout(page);